requirements.txt
streamlit
tensorflow
numpy
pandas
matplotlib
scikit-learn
seaborn
Pillow
opencv-python-headless
joblib
scipy
scikit-survival
